feat(apps): POC — same package name to eliminate the OSS import migration
Renames the OSS SDK package from @kestra-io/kestra-sdk-oss to the literal
public package name, @kestra-io/kestra-sdk, and merges its "*" dependency
range into the existing @kestra-io/kestra-sdk entry in ui/package.json
(dropping the old pinned npm dev version). npm workspaces resolves by name,
so every existing @kestra-io/kestra-sdk import in ui/src now resolves to
this local, same-commit package with zero files touched — including
stores/core.ts, reverted back to its original import path. No more
big-bang migration needed even in principle.

Also splits src/index.ts into a thin per-package wrapper plus a new
src/client-setup.ts: a createConfigureClient(client) factory generic over
any hey-api-generated client singleton, instead of closing over this
package's own generated client. This lets EE's @kestra-io/kestra-sdk
package reuse this exact file by relative import (mirroring how it already
reuses the hey-api plugin) rather than forking a second copy of the axios
setup, while keeping each package's bundled axios instance independent
(tsdown inlines this file into each package's own dist at build time).

Throwaway POC branch, not intended to ship as-is.
