+++ ui/package.json
@@ -12,7 +12,7 @@
         "build": "vite build --emptyOutDir",
         "build:design": "npm run build --workspace=@kestra-io/design-system",
         "build:topology": "npm run build --workspace=@kestra-io/topology",
-        "build:sdk-oss": "npm run build --workspace=@kestra-io/kestra-sdk-oss",
+        "build:sdk": "npm run build --workspace=@kestra-io/kestra-sdk",
         "generate:openapi-spec": "cd .. && ./gradlew :webserver:generateOpenapiSpec",
         "preview": "vite preview",
         "check:types": "vue-tsc --build packages/design-system/tsconfig.app.json && echo \"design-system compiled\" \\\n && vue-tsc --build tsconfig.app.json && echo \"app checked\" \\\n && vue-tsc --project tsconfig.test.json && echo \"test checked\"",
@@ -34,8 +34,7 @@
         "@google/genai": "^2.11.0",
         "@js-joda/core": "^6.0.1",
         "@kestra-io/design-system": "*",
-        "@kestra-io/kestra-sdk": "^2.0.0-kdevelop.17842094.03d97066.42c10b4e",
-        "@kestra-io/kestra-sdk-oss": "*",
+        "@kestra-io/kestra-sdk": "*",
         "@kestra-io/slot-contracts": "*",
         "@kestra-io/topology": "*",
         "@module-federation/enhanced": "^2.7.0",
