+++ ui/packages/kestra-sdk-oss/README.md
@@ -8,31 +8,68 @@ This package is **not published**. It exists so `ui/` can build against a locall
 generated client with no npm round-trip in the dev loop, and so a backend type change
 fails CI in the same PR that introduced it.
 
+## Same package name as the public npm package — on purpose
+
+The folder is named `kestra-sdk-oss` for humans browsing the tree, but `package.json`'s
+`name` is literally `@kestra-io/kestra-sdk` — the same name every existing import in
+`ui/src` already uses (and the same name the EE package below carries too, in its own
+repo). This is deliberate: npm workspaces resolves a bare specifier to the workspace
+member by **name**, so with the dependency range set to `"*"` in `ui/package.json`,
+every one of the ~44 existing `@kestra-io/kestra-sdk` imports in `ui/src` resolves to
+this local, in-repo package with **zero import changes** — no migration, no rename, no
+big-bang PR (the OSS/EE split reuses the exact same "same name, different local
+content" trick that already governs vs the npm registry version).
+
 ## How it's generated
 
 1. `./gradlew :webserver:generateOpenapiSpec` (run from the repo root — see the
    `generate:openapi-spec` script in `ui/package.json`) compiles `:webserver` and copies
    the resulting **OSS-only** spec to `<repo-root>/openapi.yml` (gitignored).
-2. `npm run build --workspace=@kestra-io/kestra-sdk-oss` runs `openapi-ts` (via the
-   `prebuild` hook) against that spec, then bundles with `tsdown`.
+2. `npm run build --workspace=@kestra-io/kestra-sdk` (aliased as `build:sdk`) runs
+   `openapi-ts` (via the `prebuild` hook) against that spec, then bundles with `tsdown`.
 
 ## What's hand-written vs generated
 
 Same split as the `client-sdk` repo's `javascript-sdk` package:
 
 - `src/openapi/**` — fully generated by `openapi-ts`, gitignored, never hand-edited.
-- `src/index.ts` — hand-written axios client setup (interceptors, query serializer).
+- `src/client-setup.ts` — hand-written axios client setup (interceptors, query
+  serializer, error normalization), as a `createConfigureClient(client)` **factory**
+  generic over any hey-api-generated `client` singleton — not tied to this package's own
+  generated types. This is what lets EE's `@kestra-io/kestra-sdk` package
+  (`kestra-ee/ui-ee/packages/kestra-sdk/`) reuse this exact file by relative import
+  instead of forking a second copy, each binding it to its own generated `client`. Each
+  package still bundles its own copy at build time (tsdown inlines this file into that
+  package's own `dist`), so there's no shared runtime state between the two packages —
+  only the source is shared.
+- `src/index.ts` — a thin ~5-line wrapper: re-exports the generated API, binds
+  `createConfigureClient` to this package's own `client`.
 - `heyapi-sdk-plugin/` — the custom hey-api plugin that turns the raw generated
   functions into human-friendly, tenant-aware SDK wrappers. Per the spike's requirement,
-  this is **the one copy of this plugin for the whole product**: EE's
-  `@kestra-io/kestra-sdk` package (`kestra-ee/ui-ee/packages/kestra-sdk/`) imports this
-  same module by relative path instead of forking its own copy.
+  this is **the one copy of this plugin for the whole product**: EE's package imports
+  this same module by relative path too.
+
+## The EE side needs an extra fix beyond "same name" — module resolution is per-file, not per-app
+
+Vite (like Node) resolves a bare specifier by walking up `node_modules` from the
+**real filesystem path of the importing file** — not from "which app is being built".
+`ui-ee` reuses a lot of OSS `ui/src` source directly via its `kestra` alias
+(`{find: "kestra", replacement: ".../kestra/ui/"}`), and those files still physically
+live under `kestra/ui/src/...`. Verified empirically with a minimal Vite build (see the
+kestra-ee package's README): an OSS-authored file loaded into the EE bundle this way,
+importing `@kestra-io/kestra-sdk`, resolves through **OSS's own** `node_modules` — i.e.
+this (OSS-only) package — not EE's combined one, regardless of `preserveSymlinks`. For a
+component that was written and only ever used OSS-only operations, that's actually
+correct. But it also means: don't assume "same name" alone routes every SDK call inside
+the EE bundle through EE's one configured axios instance — see
+`kestra-ee/ui-ee/vite.config.js` for the fix (an absolute-path alias, not a
+bare-specifier-to-bare-specifier one, which does not work — re-resolves relative to the
+same importer and silently falls through to the same package).
 
 ## POC scope / simplifications vs. the real `client-sdk` pipeline
 
 - No `kestra-openapi-sdk-customizer` sanitizer pass — the spec is consumed as gradle
   produces it. Reinstating that pass is straightforward follow-up work if this spike is
   adopted.
-- Only one call site (`ui/src/stores/core.ts`) was switched over to prove the mechanism
-  end-to-end. Migrating every `@kestra-io/kestra-sdk` import in `ui/src` is a separate,
-  larger effort explicitly out of scope for this spike (see its "non-goals").
+- No call site in `ui/src` needed to change (see above) — migrating away from the
+  published npm package is inherent to the mechanism itself, not a separate step.
